Export header to its own file.
