+++ Picobytes/frontend/src/pages/home/home.css
@@ -10,35 +10,4 @@
   margin-top: 70px;
 }
 
-.homepage-header {
-  position: fixed;
-  background-color: cadetblue;
-  width: 100%;
-  height: 70px;
-  top: 0;
-  left: 0;
-  z-index: 1000;
-}
-
-/* Profile icon styling */
-
-.profile-icon-container {
-  position: fixed;
-  margin-top: 10px;
-  right: 20px;
-  cursor: pointer;
-}
-
-.profile-icon {
-  border-radius: 50%;
-  min-width: 50px;
-  min-height: 50px;
-  object-fit: scale-down;
-  background-color: white;  
-}
- 
-button {
-  /* color: black; */
-  /* background-color: cadetblue; */
-}
