+++ Picobytes/frontend/src/pages/home/home_prof_overlay.tsx
@@ -2,7 +2,7 @@
 import './home_prof_overlay.css'
 import { useNavigate } from 'react-router-dom';
 
-const Homepage_Prof_Overlay = () => {
+const Home_Prof_Overlay = () => {
 
   const navigate = useNavigate();
   
@@ -22,4 +22,4 @@ const Homepage_Prof_Overlay = () => {
   );
 };
 
-export default Homepage_Prof_Overlay;
+export default Home_Prof_Overlay;
